fix: 루트 페이지의 클라이언트 버튼 분리
